Structured ArticleView Page
Added styles and code for photo caption, date, author. Need to fill out the rest of the text for the story.
Created and linked buttons under story.
Added styles and code for author information, recommended articles, and join the conversation.
Still need to add a flexbox for the author information and recommended stories.